Added title bar css
Added title bar css which includes the top buttons (currently just the
home button) and the title bar for that webpage. I made the top buttons
transparent black with the hover and non-hover having different levels
so the button is more dynamic. Also allows the button to have a similar
theme without me having to set a new color for them. The title has to
have no padding; otherwise the bar would be too wide.

I also added title css but haven't yet added it to the html. Future plan
is for the index page to have images for all the recipes.